[Hard-Fork] Implement time-activated "protocol cleanup" flag-day hard-fork, set to activate in April 2021.
To achieve desired scaling limits, the forward blocks architecture will eventually trigger a hard-fork modification of the consensus rules, for the primary purposes of dropping enforcement of many aggregate block limits and altering the difficulty adjustment algorithm.

This hard-fork will not activate until it is absolutely necessary for it to do so, at the point when real demand for additional shard space in aggregate across all forward block shard-chains exceeds the available space in the compatibility chain.  It is anticipated that this will not occur until many, many years into the future, when Freicoin/Tradecraft's usage exceeds even the levels of bitcoin usage ca. 2018.  However when it does eventually trigger, any node enforcing the old rules will be left behind.

Since the rule changes for forward blocks have not been written yet and because this flag-day code doesn't know how to detect actual activation, we cannot have older clients enforce the new rules.  What is do instead is that any rule which we anticipate changes becomes simply unenforced after this activation time, and aggregate limits are set to the maximum values the software is able to support.  After the flag-day, older clients of at least version 10.4 will continue to receive blocks, but with only SPV security ("trust the most work") for the new protocol rules.  So starting with the release of 10.4, activation of forward blocks new scaling limits becomes a soft-fork, with the only concern being the forking off of pre-10.4 nodes upon activation.

The protocol cleanup rule change is scheduled for activation on 2 April 2021 at midnight UTC.  This is 4PM PDT, 7PM EDT, and 9AM JST.  Since the activation time is median-time-past, it'll actually trigger about an hour after this wall-clock time.

This date is chosen to be roughly 2 years after the expected release date of official binaries.  While the Freicoin developer team doesn't have the resources to provide strong ongoing support beyond emergency fixes, we nevertheless have an ideal goal of supporting release binaries for up to 2 years following the first release from that series.  Any release of a new series prior to the deployment of forward blocks should set this to be at least two years from the time of that release.  When forward blocks is deployed, this parameter should be set to the highest value used in prior releases, and becomes the earliest time at which the hard-fork rules can activate.