[Chainparams] Delay IBD threshold to be 14 days instead of 24 hours.
A new node that just turns on will consider itself to be in IBD if the current time is more than a day after the latest block timestamp. This commit changes that value from 1 day to 2 weeks, as it has been known to happen that more than 24 hours pass without finding a block on the Freicoin network.