[Stratum] Allow stratum-mining activation block for block-final transactions.
The very first block after activation of block-final transactions requires the entire coinbase to be sent to a trivial / anyone-can-spend script.  To allow stratum mining to proceed past this point (e.g. for testing on a regtest network), the stratum server does not install the client's payout script if the block template assembler has altered the script in any way.