[Hard-Fork] Reduce native proof-of-work limit to 0x207fffff in protocol cleanup.
The auxiliary proof-of-work transition mechnism ensures that the native proof-of-work difficulty is reduced to a minimal value (approximately 1 diff) over the course of a few months after activation.  However the native difficulty cannot be reduced further without a hard-fork because of the minimum difficulty requirement.  This is an issue because although a diff-1 block is easy for a hardware miner to solve, but can take minutes on a CPU.